analise-sentimentos/requirements.txt
# Tested on Python 3.8.12

# Para análise de sentimentos
torch==1.10.2
torchaudio==0.10.2
torchvision==0.11.3
transformers==4.16.2

# Para acessar Kafka
kafka-python==2.0.2

# Para acessar o classificador
scikit-learn==1.0.2

# Para acessar Firebase objects
firebase-admin==5.2.0
